the site is found in the the-site folder
hey! if you find any issues, or want to give feedback, leave it on my tumblr!
https://web-based-shitposts.tumblr.com
Thank you very much for the criticism and feedback, and all that I ask you to do if you use this site (for some reason) that you credit me! thank you very much, and have a wonderful day!